Web Scraping
A company that sells books online,  It was seen that it had low sales in the “Non-Fiction” and "Travel" categories.
For this reason, it is allowed  the rival company for the web scraping.
"https://books.toscrape.com/"  in the "Travel" and "Nonfiction" category from the website taking information about books and analyzing competitors and prices needs to do.